% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffPeaks.R
\name{DiffPeaks}
\alias{DiffPeaks}
\title{This function is used to extract peak features significantly higher in treatment}
\usage{
DiffPeaks(peaks, cutoff, pcutoff, control, treat)
}
\arguments{
\item{peaks}{}

\item{cutoff}{}

\item{pcutoff}{}
}
\value{
select the samples from control and treatment groups
fold change between control and treatment
calculate pvalues
excluding error for ttest
calculating pvalues
selecting the peak features significantly higher in treatment
}
\description{
This function is used to extract peak features significantly higher in treatment
}
