% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fillpeak.R
\name{Fillpeak}
\alias{Fillpeak}
\title{Fill missing values across samples}
\usage{
Fillpeak(xset, ppm, btw, msfiles)
}
\arguments{
\item{xset}{}

\item{ppm}{}

\item{RT}{}
}
\value{
a complete peak feature matrix

peak feature matrix  
find peaks across groups  
the average mz and Rt retention time across samples    
find the peaks across samples
filling missing values  
no value to fill  
fill peaks  
extract peak intensities        
organize all peak features into a new peak matrix
}
\description{
Fill missing values across samples
}
