% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindIsotope.R
\name{FindHomo}
\alias{FindHomo}
\title{This function is used to find the primary isotopic peaks}
\usage{
FindHomo(mz1, mz2)
}
\arguments{
\item{mz1}{}

\item{mz2}{}
}
\value{

}
\description{
This function is used to find the primary isotopic peaks
}
\examples{
considering C, Cl and Br isotopes        
<0.003,H and carbon
}
