% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindIsotope.R
\name{FindIsotope}
\alias{FindIsotope}
\title{This function is used to find the primary isotopic peaks}
\usage{
FindIsotope(peaklist, mz_tol, corrcut, intensitycut)
}
\arguments{
\item{peaklist}{}

\item{mz_tol}{}

\item{corrcut}{}

\item{intensitycut}{}
}
\value{

}
\description{
This function is used to find the primary isotopic peaks
}
\examples{
the path to save results
the path to save raw data
search each mass spec files to find isotopes
no peak in this sample
search isotopes
defining searching ranges
searching the scan number for peak top  
mz ranges to find isotopes
searching isotopes using exact masses
isotopes found
only keep those mz smaller as we are interested in primary isotopic peaks
using correlations to further find isotopes
if the intensity of the primary peak is too low, peaks are excluded
calculating correlation scores
isotopes were supported by correlations
the smallest isotopic peak is the exact mass
deleting redundant isotopic peaks
deleteing redundant isotopic peaks  
}
