% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{peak_extract}
\alias{peak_extract}
\title{This function is used to extract peak features}
\usage{
peak_extract(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{

}
\description{
This function is used to extract peak features
}
