% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peakextract.R
\name{Peakextract}
\alias{Peakextract}
\title{This function is used to extract peak features}
\usage{
Peakextract(Intensitycut, ppm)
}
\arguments{
\item{Intensitycut}{}

\item{ppm}{}
}
\value{
peak features
}
\description{
This function takes raw data and then extract all peaks to dataframe
the path to save results
the path to save raw data
----------------------------------------------
detect peaks from mass spec raw files
---------------------------------------------
delete peak features with extreme m/z values
group peaks across samples
filling missing values
peak ID for each group
creating peak matrix
m/z vlaues      
retention time in minutes
replacing 0 values
extracting peak intensities across all samples
extracting peak intensities
setup the working folder
}
